refactor(eventual-send): add comments, fix style, rename index.js
