+++ packages/eventual-send/shim.js
@@ -1,5 +1,5 @@
 /* global globalThis */
-import { makeHandledPromise } from './src/index.js';
+import { makeHandledPromise } from './src/handled-promise.js';
 
 if (typeof HandledPromise === 'undefined') {
   globalThis.HandledPromise = makeHandledPromise();
