+++ packages/eventual-send/test/get-hp.js
@@ -1,5 +1,5 @@
 import makeE from '../src/E.js';
-import { makeHandledPromise } from '../src/index.js';
+import { makeHandledPromise } from '../src/handled-promise.js';
 
 export const HandledPromise = makeHandledPromise(Promise);
 export const E = makeE(HandledPromise);
